% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{admission}
\alias{admission}
\title{Graduate school admission}
\format{\code{admission} is a tbl data frame with 400 observations on 4 variables.}
\source{
\url{http://www.ats.ucla.edu/stat/r/dae/logit.htm}
}
\usage{
data(admission)
}
\description{
\code{admission} dataset contains information about how GRE (Graduate Record Exam scores), 
GPA (grade point average) and prestige of the undergraduate institution, affect admission into graduate school.
}
\details{
The 4 variables of the \code{admission} tbl data frame are the following ones:
\itemize{
\item \code{admit} admission into graduate school (binary variable: \code{0} (Not admitted) and \code{1} (Admitted))
\item \code{gre} GRE (Graduate Record Exam scores) (numeric)
\item \code{gpa} GPA (grade point average)  (numeric)
\item \code{rank} prestige of the undergraduate institution (factor with levels: \code{(1} (highest prestige), \code{(2}, \code{3} and \code{4} (lowest prestige))
}
}

