% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{awards}
\alias{awards}
\title{Students awards number}
\format{\code{awards} is a tbl data frame with 200 observations on 4 variables.}
\source{
\url{http://www.ats.ucla.edu/stat/r/dae/poissonreg.htm}
}
\usage{
data(awards)
}
\description{
\code{awards} dataset contains information about the number of awards earned by students at one high school, 
the type of program in which the student was enrolled (e.g., vocational, general or academic) and the score on their final exam in math.
}
\details{
The 4 variables of the \code{awards} tbl data frame are the following ones:
\itemize{
\item \code{id} students id (factor)
\item \code{num_awrds} number of awards earned by students at a high school in a year (numeric)
\item \code{prog} students' scores on their math final exam (numeric)
\item \code{math}  type of program in which the students were enrolled (factor with levels: \code{(1} (General), \code{(2} (Academic) and \code{3} (Vocational))
}
}

