% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{bands}
\alias{bands}
\title{Cylinder Bands}
\format{\code{bands} is a tbl data frame with 540 observations on 40 variables.}
\source{
This dataset is publicly available at \href{http://archive.ics.uci.edu/ml/datasets/Cylinder+Bands}{UCI Machine Learning repository}.

Please include this citation if you plan to use this dataset:\cr
Lichman, M. (2013). UCI Machine Learning Repository [\url{http://archive.ics.uci.edu/ml}]. Irvine, CA: University of California, School of Information and Computer Science.

The dataset was created and donated on 08 Jan 1995 to \href{http://archive.ics.uci.edu/ml/datasets/Cylinder+Bands}{UCI Machine Learning repository} by:\cr
Bob Evans 
RR Donnelley & Sons Co. 
Gallatin Division 
801 Steam Plant Rd 
Gallatin, Tennessee 37066-3396 
(615) 452-5170
}
\usage{
data(bands)
}
\description{
This dataset is related with process delays known as cylinder banding in rotogravure printing.

The full dataset was described and analyzed in [Moro et al., 2011]., see \code{source} section.
}
\details{
The 40 variables are organized as follows.

\itemize{
\item \code{timestamp} (numeric from 19900330 to 19941010);
\item \code{cylinder_number} (character);
\item \code{customer} (character);
\item \code{job_number} (numeric from 23040 to 88231) 
\item \code{grain_screened} (factor with 2 levels: YES, NO);
\item \code{ink_color} (factor with 1 level: KEY);
\item \code{proof_on_ctd_ink} (factor with 2 levels: YES, NO);
\item \code{blade_mfg} (factor with 2 levels: BENTON, UDDEHOLM);
\item \code{cylinder_division} (factor with 1 level: GALLATIN);
\item \code{paper_type} (factor with 3 level: UNCOATED, COATED, SUPER);
\item \code{ink_type} (factor with 3 level: UNCOATED, COATED, COVER); 
\item \code{direct_steam} (factor with 2 levels: YES, NO);
\item \code{solvent_type} (factor with 3 level: LINE, NAPTHA, XYLOL);
\item \code{type_on_cylinder} (factor with 2 levels: YES, NO);
\item \code{press_type} (factor with 4 levels: ALBERT70, MOTTER70, MOTTER94, WOODHOE70);
\item \code{press} (numeric from 802 to 828)
\item \code{unit_number}: printing unit (numeric: 1-10)
\item \code{cylinder_size} (factor with 3 levels: CATALOG, SPIEGEL, TABLOID);
\item \code{paper_mill_location} (factor with 5 levels: CANADIAN, MIDEUROPEAN, NORTHUS, SCANDANAVIAN, SOUTHUS);
\item \code{plating_tank} (numeric: 1910, 1911);
\item \code{proof_cut} measured in percentage (numeric: 0-100);
\item \code{viscosity} measured in percentage (numeric: 0-100);
\item \code{caliper} measured in percentage (numeric: 0-1);
\item \code{ink_temperature} measured in degrees Fahrenheit (numeric: 5-30);
\item \code{humidity} measured in percentage (numeric: 5-105);
\item \code{roughness} measured in microns (numeric: 0-2);
\item \code{blade_pressure} measured in pounds (numeric: 10-75);
\item \code{varnish_pct} measured in percentage (numeric: 0-100);
\item \code{press_speed} measured in feet per minute (numeric: 0-4000);
\item \code{ink_pct} measured in percentage (numeric: 0-100);
\item \code{solvent_pct} measured in percentage (numeric: 0-100);
\item \code{esa_voltage} measured in kilovolts (numeric: 0-16);
\item \code{esa_amperage} measured in milliamps (numeric: 0-10);
\item \code{wax} measured in gallons (numeric: 0-4);
\item \code{hardener} measured in gallons (numeric: 0-3.0);
\item \code{roller_durometer} (numeric: 15-120);
\item \code{current_density} measured in amperes per square decimeter (numeric: 20-50);
\item \code{anode_space_ratio} (numeric: 70-130);
\item \code{chrome_content} (numeric: 80-120);
\item \code{band_type} (factor with 2 levels: BAND, NOBAND).
}
}

