% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{bank}
\alias{bank}
\title{Marketing Campaigns of a Portuguese Banking Institution}
\format{\code{bank} is a tbl data frame with 45211 observations on 20 variables.

The data is ordered by call id (and also by date, from May 2008 to November 2010).}
\source{
This dataset is publicly available for research at \href{http://archive.ics.uci.edu/ml/datasets/Bank+Marketing}{UCI Machine Learning repository}.
The details are described in [Moro et al., 2011]. 

Please include this citation if you plan to use this dataset:\cr
[Moro et al., 2011] S. Moro, R. Laureano and P. Cortez. Using Data Mining for Bank Direct Marketing: An Application of the CRISP-DM Methodology.  
In P. Novais et al. (Eds.), \emph{Proceedings of the European Simulation and Modelling Conference - ESM'2011}, pp. 117-121, Guimarães, Portugal, October, 2011. EUROSIS.\cr
Available at \url{http://hdl.handle.net/1822/14838}.
}
\usage{
data(bank)
}
\description{
This dataset is related with direct marketing campaigns of a Portuguese
banking institution. The marketing campaigns were based on phone calls. 
Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed. 

The full dataset was described and analyzed in [Moro et al., 2011]., see \code{source} section.
}
\details{
The 20 variables are organized as follows.

Bank client data:
\itemize{
\item \code{id} phone call id (integer: from 1 to 45211);
\item \code{age} age (integer);
\item \code{job} type of job (factor with 12 levels: admin., unknown, unemployed, management, housemaid, entrepreneur, student,  blue-collar, self-employed, retired, technician, services);
\item \code{marital} marital status (factor with 3 levels: married, divorced, single). divorced means divorced or widowed;
\item \code{education} education (factor with 4 levels: unknown, secondary, primary, tertiary);
\item \code{default} has credit in default? (factor with 2 levels: yes, no);
\item \code{balance} average yearly balance, in euros (integer);
\item \code{housing} has housing loan? (factor with 2 levels: yes, no);
\item \code{loan} has personal loan? (factor with 2 levels: yes, no).
}
Related with the last contact of the current campaign:
\itemize{
\item \code{contact} contact communication type (factor with 3 levels: unknown, telephone, cellular);
\item \code{day} last contact day of the month (integer);
\item \code{month} last contact month of year (factor with 12 levels: jan, feb, mar, ..., nov, dec);
\item \code{year} last contact year (integer: 2008, 2009, 2010);
\item \code{date} last contact date (\code{POSIXct} date);
\item \code{duration} last contact duration, in seconds (integer);
\item \code{y} has the client subscribed a term deposit? (factor with 2 levels: yes, no).
}
Other attributes:
\itemize{
\item \code{campaign} number of contacts performed during this campaign and for this client (integer: it includes last contact);
\item \code{pdays} number of days that passed by after the client was last contacted from a previous campaign (integer). -1 means client who was not previously contacted;
\item \code{previous} number of contacts performed before this campaign and for this client (integer);
\item \code{poutcome} outcome of the previous marketing campaign (factor with 4 levels: "unknown", "other", "failure", "success").
}
}

