% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{banknotes}
\alias{banknotes}
\title{Swiss Banknotes}
\format{\code{banknotes} is a tbl data frame with 200 observations on 7 variables.}
\source{
Härdle and Simar, Applied Multivariate Statistical Analysis, 3rd ed., Springer, 2012
}
\usage{
data(banknotes)
}
\description{
This dataset deals with 200 old Swiss 1,000-franc banknotes classified by dimensions (6 numerical characteristics) and type (genuine/counterfeit).
The first 100 banknotes are genuine, the last 100 are counterfeit.
}
\details{
The 7 variables of the \code{banknotes} tbl data frame are the following ones:
\itemize{
\item \code{length} length of the banknote (numeric);
\item \code{left} height of the banknote, measured on the left (cu.in.) (numeric);
\item \code{right} height of the banknote, measured on the right (numeric);
\item \code{bottom} distance of inner frame to the lower border (numeric);
\item \code{top} distance of inner frame to the upper border (numeric);
\item \code{diagonal} length of the diagonal (numeric);
\item \code{type} if the banknote is genuine or counterfeit (Factor).
}
}

