% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{bimetal1}
\alias{bimetal1}
\title{Bimetal Dataset in Phase I}
\format{\code{bimetal1} is a tbl data frame with 28 observations on 5 variables.}
\usage{
data(bimetal1)
}
\description{
The \code{bimetal1} data was extracted from the analogous data frame of package \code{MSQC}. 
The dataset contains measurements of the deflection, the curvature, the resistivity and the hardness in low 
and high expansion sides, from brass and steel bimetal thermostats. The manufacturing process is assumed to be "in-control".
}
\details{
The 5 variables of the \code{bimetal1} tbl data frame are the following ones:
\itemize{
\item \code{deflection} the deflection level in 10^-6 1/Kl (numeric);
\item \code{curvature} the curvature level in 10^-6 1/K (numeric);
\item \code{resistivity} the resistivity level in 10 ^-1ohm x mm^2/m (numeric);
\item \code{Hardness.low.side} the hardness of the low expansion side in 10 N/mm^2 (numeric);
\item \code{Hardness.high.side} the hardness of the high expansion side in 10 N/mm^3 (numeric).
}
}
\seealso{
the complete \code{bimetal1} help page is in the \code{MSQC} package: \code{\link[MSQC]{bimetal1}}
}

