% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{boiling}
\alias{boiling}
\title{Boiling Dataset}
\format{\code{boiling} is a tbl data frame with 6 observations on 3 variables.}
\usage{
data(boiling)
}
\description{
In an attempt to resolve a domestic dispute about which of two pans was the quicker pan for cooking, the following data were obtained.
Various measured volumes (pints) of cold water were put into each pan and geated using the same setting of the cooker.
The response variable was the `time` in minutes until the water boiled.
}
\details{
The 3 variables of the \code{boiling} tbl data frame are the following ones:
\itemize{
\item \code{time} time to boil (minutes) (numeric);
\item \code{pan} type of pan (factor).
\item \code{volume} volume of water (pints) (numeric).
}
}

