% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{bostonhousing}
\alias{bostonhousing}
\title{Boston Houses}
\format{\code{bostonhousing} is a tbl data frame with 506 observations on 14 variables.}
\usage{
data(bostonhousing)
}
\description{
The \code{bostonhousing} data deals with 506 houses in Boston classified by 13 numerical characteristics
(age, tax, etc.) and one categorical characteristic (chas).
}
\details{
The 14 variables of the \code{bostonhousing} tbl data frame are the following ones:
\itemize{
\item \code{crim} per capita crime rate by town (numeric)
\item \code{zn} proportion of residential land zoned for lots over 25,000 sq.ft. (numeric)
\item \code{indus} proportion of non-retail business acres per town (numeric)
\item \code{chas} Charles River dummy variable (= 1 if tract bounds river; 0 otherwise) (numeric)
\item \code{nox} nitrogen oxides concentration (parts per 10 million) (numeric)
\item \code{rm} average number of rooms per dwelling (numeric)
\item \code{age} proportion of owner-occupied units built prior to 1940 (numeric)
\item \code{dis} weighted mean of distances to five Boston employment centres (numeric)
\item \code{rad} index of accessibility to radial highways (numeric)
\item \code{tax} full-value property-tax rate per $10,000 (numeric)
\item \code{ptratio} pupil-teacher ratio by town (numeric)
\item \code{black} 1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town (numeric)
\item \code{lstat} lower status of the population (percent)  (numeric)
\item \code{medv} median value of owner-occupied homes in $1000 (numeric)
}
}

