% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{bottlecap}
\alias{bottlecap}
\title{Mean diameter of caps produced by a forging machine}
\format{\code{bottlecap} is a tbl data frame with 840 observations on 8 variables.}
\usage{
data(bottlecap)
}
\description{
\code{bottlecap} dataset contains measures of the mean diameter of the caps produced by a forging machine
}
\details{
The 8 variables of the \code{bottlecap} tbl data frame are the following ones:
\itemize{
\item \code{Cavity.1} mean diameter measures of cap produced by cavity 1 (numeric)
\item \code{Cavity.2} mean diameter measures of cap produced by cavity 2 (numeric)
\item \code{Cavity.3} mean diameter measures of cap produced by cavity 3 (numeric)
\item \code{Cavity.4} mean diameter measures of cap produced by cavity 4 (numeric)
\item \code{Cavity.5} mean diameter measures of cap produced by cavity 5 (numeric)
\item \code{Cavity.6} mean diameter measures of cap produced by cavity 6 (numeric)
\item \code{Cavity.7} mean diameter measures of cap produced by cavity 7 (numeric)
\item \code{Cavity.8} mean diameter measures of cap produced by cavity 8 (numeric)
}
}

