% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{brainbod}
\alias{brainbod}
\title{Body weight and brain weight}
\format{\code{brainbod} is a tbl data frame with 15 observations on 3 variables.}
\usage{
data(brainbod)
}
\description{
\code{brainbod} dataset contains information about the weight of body and brain of different species of animals.
}
\details{
The 3 variables of the \code{brainbod} tbl data frame are the following ones:
\itemize{
\item \code{Species} animal species (character)
\item \code{Body} weight of the animal (in kg) (numeric)
\item \code{Brain} weight of the brain of the animal (in g) (numeric)
}
}

