% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{bwt}
\alias{bwt}
\title{Low Birthweight}
\format{\code{bwt} is a tbl data frame with 189 observations on 4 variables.}
\usage{
data(bwt)
}
\description{
The low birthweight data contains information on low birth weight in infants born at a US hospital. 
A number of variables were collected that might explain the cause of the low birth weight.
}
\details{
The 7 variables of the \code{bwt} tbl data frame are the following ones:
\itemize{
\item \code{low}  indicator of birth weight less than 2.5 kg. (numeric)
\item \code{age}  mother's age in years (numeric)
\item \code{lwt}  mother's weight in pounds at last menstrual period (numeric)
\item \code{smoke} smoking status during pregnancy (factor)
}
}

