% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{carctl}
\alias{carctl}
\title{Parking time}
\format{\code{carctl} is a tbl data frame with 20 observations on 2 variables.}
\usage{
data(carctl)
}
\description{
\code{carctl} data contains information of an experiment which investigates if an optional equipment of cars may help to park the cars themselves. 
So, 20 drivers were asked to park the car with and without the equipment. 
The parking time (secs) for each driver and equipment is recorded.
}
\details{
The 2 variables of the \code{carctl} tbl data frame are the following ones:
\itemize{
\item \code{Car_a} parking time of car with equipment a (seconds) (numeric)
\item \code{Car_b} parking time of car with equipment b (seconds) (numeric)
}
}

