% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{carseat}
\alias{carseat}
\title{Quality of car seats}
\format{\code{carseat} is a tbl data frame with 75 observations on 2 variables.}
\usage{
data(carseat)
}
\description{
\code{carseat} contains data of an experiment about the reproducibility of a measurement method aiming to test resistance of a specific material used to cover car seats.
As a result, an experiments involving 75 samples of material from the same batch is set up. Three operators: Kevin, Michelle and Rob are assigned to test 25 samples each.
}
\details{
The 2 variables of the \code{carseat} tbl data frame are the following ones:
\itemize{
\item \code{Operator} operator which performed the measurement (factor with levels: \code{Michelle}, \code{Kevin} and \code{Rob})
\item \code{Strength} breaking load of the fabric (in kg) (numeric)
}
}

