% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{cerealbx}
\alias{cerealbx}
\title{Filling cereal boxes}
\format{\code{cerealbx} is a tbl data frame with 6 observations on 1 variables.}
\usage{
data(cerealbx)
}
\description{
\code{cerealbx} dataset contains data on the filling of cereal boxes.
}
\details{
The variable of the \code{cerealbx} tbl data frame is:
\itemize{
\item \code{BoxWeigh} weight of cereal boxes (g) (numeric)
}
}

