% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{clotting}
\alias{clotting}
\title{Clotting Dataset}
\format{\code{clotting} is a tbl data frame with 18 observations on 3 variables.}
\usage{
data(clotting)
}
\description{
From McCullagh and Nelder (1989, pp. 300-302). Hurn et al. (1945) published data on the clotting time of blood, 
giving clotting times in seconds (y) for normal plasma diluted to nine different percentage concentration with prothrombin-free
plasma (u); clotting was induce by two lots of thromboplastin. Data are analyzed using gamma errors and inverse link.
}
\details{
The 3 variables of the \code{clotting} tbl data frame are the following ones:
\itemize{
\item \code{u} percentages of diluition (numeric).
\item \code{y} clotting time (sec) (numeric);
\item \code{lot} lot of thromboplastin (factor).
}
}

