% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{default}
\alias{default}
\title{Default of credit card clients}
\format{\code{default} is a tbl data frame with 30000 observations on 25 variables.}
\source{
I-Cheng Yeh, Department of Information Management, Chung Hua University, Taiwan. (2) Department of Civil Engineering, Tamkang University, Taiwan.
}
\usage{
data(default)
}
\description{
The \code{default} data contains information about customers default payments in Taiwan.
}
\details{
The 25 variables of the \code{default} tbl data frame are the following ones:
\itemize{
\item \code{ID} (numeric);
\item \code{LIMIT_BAL} Amount of the given credit (NT dollar): it includes both the individual consumer credit and his/her family (supplementary) credit (numeric)
\item \code{SEX} Gender (factor with 2 levels: \code{1} (Male) and \code{2}(Female));
\item \code{EDUCATION} Education (factor with 7 levels: \code{0}, \code{1}, \code{2}, \code{3}, \code{4}, \code{5} and \code{6});
\item \code{MARRIAGE} Marital status (factor with 4 levels: \code{0}, \code{1}, \code{2} and \code{3});
\item \code{AGE} Age (year) (numeric);
\item \code{PAY_0} History of past payment: the repayment status in September, 2005 (numeric);
\item \code{PAY_2} History of past payment: the repayment status in August, 2005 (numeric);
\item \code{PAY_3} History of past payment: the repayment status in July, 2005 (numeric);
\item \code{PAY_4} History of past payment: the repayment status in June, 2005 (numeric);
\item \code{PAY_5} History of past payment: the repayment status in May, 2005 (numeric);
\item \code{PAY_6} History of past payment: the repayment status in April, 2005 (numeric);
\item \code{BILL_AMT1} Amount of bill statement (NT dollar): amount of bill statement in September, 2005 (numeric);
\item \code{BILL_AMT2} Amount of bill statement (NT dollar): amount of bill statement in August, 2005(numeric);
\item \code{BILL_AMT3} Amount of bill statement (NT dollar): amount of bill statement in July, 2005 (numeric);
\item \code{BILL_AMT4} Amount of bill statement (NT dollar): amount of bill statement in June, 2005 (numeric);
\item \code{BILL_AMT5} Amount of bill statement (NT dollar): amount of bill statement in May, 2005 (numeric);
\item \code{BILL_AMT6} Amount of bill statement (NT dollar): amount of bill statement in April, 2005 (numeric);
\item \code{PAY_AMT1} Amount of previous payment (NT dollar): amount paid in September (numeric);
\item \code{PAY_AMT2} Amount of previous payment (NT dollar): amount paid in August (numeric);
\item \code{PAY_AMT3} Amount of previous payment (NT dollar): amount paid in July (numeric);
\item \code{PAY_AMT4} Amount of previous payment (NT dollar): amount paid in June (numeric);
\item \code{PAY_AMT5} Amount of previous payment (NT dollar): amount paid in May (numeric);
\item \code{PAY_AMT6} Amount of previous payment (NT dollar): amount paid in April (numeric);
\item \code{default_payment} default payment next month (factor with 2 levels: \code{0} (No) and \code{1} (Yes));
}
}

