% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{diabetes}
\alias{diabetes}
\title{Diabetes Patients Records}
\format{\code{diabetes} is a tbl data frame with 30264 observations on 5 variables.}
\source{
This dataset is publicly available at \href{http://archive.ics.uci.edu/ml/datasets/Diabetes}{UCI Machine Learning repository}.
}
\usage{
data(diabetes)
}
\description{
Insulin Dependent Diabetes Mellitus (IDDM) data for the AAAI Spring Symposium (1994).
}
\details{
Diabetes patient records were obtained from two sources: an automatic electronic recording device and paper records. The automatic device had an internal clock to timestamp events, whereas the paper records only provided "logical time" slots (breakfast, lunch, dinner, bedtime).  For paper records, fixed times were assigned to breakfast (08:00), lunch (12:00), dinner (18:00), and bedtime (22:00).  Thus paper records have fictitious uniform recording times whereas electronic records have more realistic time stamps. 

Data has five variables:
\itemize{
\item \code{patient} id (character), this correspond to a single file on original data
\item \code{date} in MM-DD-YYYY format (factor)
\item \code{time} in XX:YY format (factor)
\item \code{code} indicates a task, according to the following list (integer)
\itemize{  
\item 33 = Regular insulin dose
\item 34 = NPH insulin dose
\item 35 = UltraLente insulin dose
\item 48 = Unspecified blood glucose measurement
\item 57 = Unspecified blood glucose measurement
\item 58 = Pre-breakfast blood glucose measurement
\item 59 = Post-breakfast blood glucose measurement
\item 60 = Pre-lunch blood glucose measurement
\item 61 = Post-lunch blood glucose measurement
\item 62 = Pre-supper blood glucose measurement
\item 63 = Post-supper blood glucose measurement
\item 64 = Pre-snack blood glucose measurement
\item 65 = Hypoglycemic symptoms
\item 66 = Typical meal ingestion
\item 67 = More-than-usual meal ingestion
\item 68 = Less-than-usual meal ingestion
\item 69 = Typical exercise activity
\item 70 = More-than-usual exercise activity
\item 71 = Less-than-usual exercise activity
\item 72 = Unspecified special event
}
\item \code{value} insuline value (character)
}
}

