% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{diameters}
\alias{diameters}
\title{Pin diameters}
\format{\code{diameters} is a tbl data frame with 50 observations on 3 variables.}
\usage{
data(diameters)
}
\description{
\code{diameters} data contains data collected from five different lathes, each of these used by two different operators.
}
\details{
The 3 variables of the \code{diameters} tbl data frame are the following ones:
\itemize{
\item \code{Lathe} lathe (factor with levels \code{1}, \code{2}, \code{3}, \code{4}, \code{5})
\item \code{Operator} operator (factor with levels \code{D} and \code{N})
\item \code{Pin.Diam} pin diameter measure (numeric)
}
}

