% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{dicentric}
\alias{dicentric}
\title{Chromosomal Abnormalities Data}
\format{\code{dicentric} is a tbl data frame with 27 observations on 4 variables.}
\usage{
data(dicentric)
}
\description{
\code{dicentric} data contains numbers (\code{ca}) of chromosomal abnormalities observed for various amounts (\code{doseamt})
and intensities (\code{doserate}) of gamma radiation. The numbers of \code{cells} per measurement varied.
}
\details{
The 4 variables of the \code{dicentric} tbl data frame are the following ones:
\itemize{
\item \code{cells} Number of cells analyzed (numeric)
\item \code{ca}  Number of chromosomal abnormalities read (numeric)
\item \code{doseamt} Amount of gamma radiation (numeric)
\item \code{doserate} Dose rate of gamma radiation (numeric)
}
}

