% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{drug}
\alias{drug}
\title{Reaction Times}
\format{\code{drug} is a tbl data frame with 10 observations on 3 variables.}
\usage{
data(drug)
}
\description{
\code{drug} dataset contains the results of an experiment to investigate the effects of a depressant drug. In particular,
 the reaction times of ten male rats to a certain stimulus were measured after a specified dose of the drug have been administered to each rat.
}
\details{
The 3 variables of the \code{drug} tbl data frame are the following ones:
\itemize{
\item \code{rat} rat id (numeric);
\item \code{dose} specified dose of the drug administered (numeric);
\item \code{time} reaction time (numeric).
}
}

