% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{druguse}
\alias{druguse}
\title{Drug Rates Correlations}
\format{\code{druguse} is a 13 x 13 correlation matrix.}
\source{
Everitt, B. and Hothorn, T., *An Introduction to Applied Multivariate Analysis with `R`*, Springer, 2011
}
\usage{
data(druguse)
}
\description{
The \code{druguse} data contains the correlation matrix among rates (recorded on a 5-point scale)
for a sample of 1,634 students in the seventh to ninth grades in 11 schools in the greater metropolitan area
of Los Angeles. Each participant completed a questionnaire about the number of times a particular substance had ever been used.
}
\details{
The 13 substances asked about, which are also the row and columns names of the matrix, were the following ones:
\itemize{
\item \code{cigarettes};
\item \code{beer};
\item \code{wine};
\item \code{liquor};
\item \code{cocaine};
\item \code{tranquillizers};
\item \code{drug store medications} used to get high;
\item \code{heroin} and other opiates;
\item \code{marijuana};
\item \code{hashish};
\item \code{inhalants} (glue, gasoline, etc.);
\item \code{hallucinogenics} (LSD, mescaline, etc.);
\item \code{amphetamine} stimulants.
}
}

