% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{ds}
\alias{ds}
\title{Measures of weigth and heigth in a sample of young people}
\format{\code{ds} is a tbl data frame with 200 observations on 3 variables.}
\usage{
data(ds)
}
\description{
\code{ds} data contains information about weigth and height of a sample of young people. Also gender is recorded.
}
\details{
The 3 variables of the \code{ds} tbl data frame are the following ones:
\itemize{
\item \code{height}  height (in cm) (numeric)
\item \code{weight}  weight (in Kg) (numeric)
\item \code{gender}  gender of a different young person (factor with levels: \code{M} and \code{F})
}
}

