% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{forbes94}
\alias{forbes94}
\title{Most Paid CEOs in U.S.}
\format{\code{forbes94} is a tbl data frame with 800 observations on 30 variables.}
\usage{
data(forbes94)
}
\description{
The \code{forbes94} data deals with the compensation received in 1994 by the 800 most paid CEOs in US
(This dataset comes from a ranking published by the Forbes newspaper in the same year).
}
\details{
The 30 variables of the \code{forbes94} tbl data frame are the following ones:
\itemize{
\item \code{TotalComp} (numeric)
\item \code{WideIndustry} Factor w/ 19 levels
\item \code{Company} Factor w/ 800 levels
\item \code{CEO} Factor w/ 796
\item \code{CityofBirth} Factor w/ 402 levels
\item \code{StateofBirth} Factor w/ 51 levels
\item \code{Age} (numeric)
\item \code{Undergrad} Factor w/ 262 levels
\item \code{UGDegree} Factor w/ 25 levels 
\item \code{UGDate} (numeric)
\item \code{AgeOfUnder} (numeric)
\item \code{Graduate}Factor w/ 129 levels
\item \code{GradDegree} Factor w/ 20 levels
\item \code{MBA} (numeric)
\item \code{MasterPhd} (numeric)
\item \code{G_date} (numeric)
\item \code{AgeOfGradu} (numeric)
\item \code{YearsFirm} (numeric)
\item \code{YearsCEO} (numeric)
\item \code{Salary} (numeric)
\item \code{Bonus} (numeric)
\item \code{Other} (numeric)
\item \code{StGains} (numeric)
\item \code{Compfor5Yrs} (numeric)
\item \code{StockOwned} (numeric)
\item \code{Sales} (numeric)
\item \code{Profits} (numeric)
\item \code{ReturnOver5Yrs} (numeric)
\item \code{Industry} (numeric)
\item \code{IndustryCode} (numeric)
}
}

