% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{g1res}
\alias{g1res}
\title{G1res}
\format{\code{g1res} is a tbl data frame with 19 observations on 3 variables.}
\usage{
data(g1res)
}
\description{
The \code{g1res} data was ......
}
\details{
The 3 variables of the \code{g1res} tbl data frame are the following ones:
\itemize{
\item \code{V1} (numeric);
\item \code{V2} (numeric);
\item \code{V3} (numeric).
}
}

