% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{german}
\alias{german}
\alias{germannumeric}
\alias{germanstring}
\title{German Credit Cards}
\format{\code{german} is a tbl data frame with 1000 observations on 21 variables.}
\usage{
data(german)
}
\description{
The \code{german} data deals with a sample of 1,000 customers which apply for a credit card. Some have defaulted
(Bad class) and other not (Good) on the credit card. 20 further characteristics about them or their credit card have been measured.
}
\details{
The 21 variables of the \code{german} tbl data frame are the following ones:
\itemize{
\item \code{StatusAccount} Factor w/ 4 levels
\item \code{Duration} (numeric)
\item \code{CreditHistory} Factor w/ 5 levels
\item \code{Purpose} Factor w/ 10 levels
\item \code{CreditAmount} (numeric)
\item \code{SavingsAccountBonds} Factor w/ 5 levels
\item \code{EmployedSince} Factor w/ 5 levels
\item \code{RatePercIncome} (numeric)
\item \code{StatusAndSex} Factor w/ 5 levels
\item \code{OtherBebtorsGarants} Factor w/ 3 levels
\item \code{ResidenceSince} (numeric)
\item \code{Property} Factor w/ 4 levels
\item \code{AgeYears} (numeric)
\item \code{OtherInstallPlans} Factor w/ 3 levels
\item \code{Housing} Factor w/ 3 levels
\item \code{NumberOfCredits} (numeric)
\item \code{Job} Factor w/ 4 levels
\item \code{NumberPeopleMaintained} (numeric)
\item \code{Telephone} Factor w/ 2 levels
\item \code{ForeignWorker} Factor w/ 2 levels
\item \code{Class} Factor w/ 2 levels
}
}

