% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{hormones}
\alias{hormones}
\title{Growth hormones}
\format{\code{hormones} is a tbl data frame with 18 observations on 2 variables.}
\usage{
data(hormones)
}
\description{
\code{hormones} data contains the results of an experiment to compare two growth hormones. 
18 chicken were tested: 10 being assigned to hormone A and 8 to hormone B. The gains in weights (grams) over the period of experiment were measured.
}
\details{
The 2 variables of the \code{hormones} tbl data frame are the following ones:
\itemize{
\item \code{hormone} growth hormone given to chicken (factor with levels: \code{A} and \code{B})
\item \code{gain}  gain in weigth (grams) (numeric)
}
}

