% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{hotdogs}
\alias{hotdogs}
\title{Hot Dogs}
\format{\code{hotdogs} is a tbl data frame with 54 observations on 3 variables.}
\usage{
data(hotdogs)
}
\description{
\code{hotdogs} data contains calories and sodium content for three types of hot dogs: 
20 Beef, 17 Meat and 17 Poultry hot dogs. This data is a sample from much larger populations.
}
\details{
The 3 variables of the \code{hotdogs} tbl data frame are the following ones:
\itemize{
\item \code{type} hotdog type (factor with levels: \code{beef}, \code{meat} and \code{poultry})
\item \code{calories} calories content in Kcal (numeric)
\item \code{sodium} sodium content (numeric)
}
}

