% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{housing}
\alias{housing}
\title{Frequency Table from a Copenhagen Housing Conditions Survey}
\format{\code{housing} is a tbl data frame with 72 observations on 5 variables.}
\source{
The complete \code{housing} help page is in the \code{MASS} package: \code{\link[MASS]{housing}}
}
\usage{
data(housing)
}
\description{
\code{housing} contains information on housing conditions in Copenhagen of 1681 householders who were surveyed on: 
the type of rental accommodation they occupied, the degree of contact they had with other residents, their feeling of influence on apartment management,
and on their level of satisfaction with their housing conditions.
}
\details{
The 5 variables of the \code{housing} tbl data frame are the following ones:
\itemize{
\item \code{Sat} Satisfaction of householders with their present housing circumstances (factor with levels: \code{High}, \code{Medium} and \code{Low})
\item \code{Infl} Perceived degree of influence householders have on the management of the property (factor with levels: \code{High}, \code{Medium} and \code{Low})
\item \code{Type} Type of rental accommodation (factor with levels: \code{Tower}, \code{Atrium}, \code{Apartment} and \code{Terrace})
\item \code{Cont} Contact residents are afforded with other residents (factor with levels: \code{(Low} and \code{High})
\item \code{Freq} Frequencies: the numbers of residents in each class (numeric)
}
}

