% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{iowheat}
\alias{iowheat}
\title{Yield of Wheat}
\format{\code{iowheat} is a tbl data frame with 33 observations on 10 variables.}
\usage{
data(iowheat)
}
\description{
\code{iowheat} dataset is a toy example that summarizes the yield of wheat (bushels per acre) 
for the state of Iowa between 1930-1962. In addition to yield, year, rainfall and temperature were recorded 
as the main predictors of yield.
}
\details{
The 10 variables of the \code{iowheat} tbl data frame are the following ones:
\itemize{
\item \code{Year} year of harvest (numeric)
\item \code{Rain0} pre-season rainfall (numeric)
\item \code{Temp1} mean temperature for growing month 1 (numeric)
\item \code{Rain1} rainfall for growing month 1 (numeric)
\item \code{Temp2} mean temperature for growing month 2 (numeric)
\item \code{Rain2} rainfall for growing month 2 (numeric)
\item \code{Temp3} mean temperature for growing month 3 (numeric)
\item \code{Rain3} rainfall for growing month 3 (numeric)
\item \code{Temp4} mean temperature for harvest month (numeric)
\item \code{Yield} yield in bushels per acre (numeric)
}
}

