% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{irished}
\alias{irished}
\title{The Raftery and Hout Irish education data}
\format{\code{irished} is a tbl data frame with 500 observations on 6 variables.}
\source{
Data collected by Greaney and Kelleghan (1984), and reanalyzed by Raftery and Hout (1985, 1993)
}
\usage{
data(irished)
}
\description{
\code{irished} data contains Data on educational transitions for a sample of 500 Irish schoolchildren aged 11 in 1967.
}
\details{
The 6 variables of the \code{irished} tbl data frame are the following ones:
\itemize{
\item \code{Sex} Factor with levels 1 = male; 2 = female
\item \code{DVRT} Drumcondra Verbal Reasoning Test Score (numeric)
\item \code{edlevel} Educational level attained:
\enumerate{
\item Primary terminal leaver
\item Junior cycle incomplete: vocational school
\item Junior cycle incomplete: secondary school
\item Junior cycle terminal leaver: vocational school
\item Junior cycle terminal leaver: secondary school
\item Senior cycle incomplete: vocational school
\item Senior cycle incomplete: secondary school
\item Senior cycle terminal leaver: vocational school
\item Senior cycle terminal leaver: secondary school
\item 3rd level incomplete
\item 3rd level complete
 }
\item \code{lvcert} Leaving Certificate. Factor with 2 levels: 1 if Leaving Certificate not taken; 2 if taken.
\item \code{fathocc} Prestige score for father's occupation (calculated by Raftery and Hout, 1985) (numeric)
\item \code{schltype} Type of school. Factor of 3 levels: 1 = secondary; 2 = vocational; 9 = primary terminal leaver.
}
}

