% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{istat}
\alias{istat}
\title{People Features}
\format{\code{istat} is a tbl data frame with 1806 observations on 4 variables.}
\usage{
data(istat)
}
\description{
\code{istat} data contain weight, height, gender and geographical area 
("Nord", "Centro", "Sud" and "Isole") from 1806 Italian people.
}
\details{
The 4 variables of the \code{istat} tbl data frame are the following ones:
\itemize{
\item \code{Gender} gender (factor with levels: \code{Female} and \code{Male})
\item \code{Area} geographical area (factor with levels: \code{Nord}, \code{Centro}, \code{Sud} and \code{Isole})
\item \code{Weight} weight in kg (numeric)
\item \code{Height} height in cm (numeric)
}
}

