% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{italia}
\alias{italia}
\title{Data on Italian Districts}
\format{\code{italia} is a tbl data frame with 8003 observations on 14 variables.}
\source{
This dataset was downloaded in the Istat website: \href{http://www.istat.it/it/archivio/6789}{Istituto nazionale di statistica (Istat)}.
}
\usage{
data(italia)
}
\description{
The \code{italia} dataset is related with information on the 8003 Italian districts ("comuni") updated at 2016-01-01. 
In particular information on "province", "regioni, "ripartizioni" ("major" Italian districts) related to "comuni" is available.
}
\details{
The 14 variables of the \code{italia} tbl data frame are the following ones:
\itemize{
\item \code{cod_comune} codes of the Italian "comuni" (character);
\item \code{comune} names of the Italian "comuni" (character);
\item \code{comune_cap_prov} if a "comune" is the administrative center of the "provincia" (logical);
\item \code{pop_legale} legal population living in the "comuni" at the last "Censimento" (2011-10-09) (integer);
\item \code{progr_comune} progressive codes of the Italian "comuni" (it restarts from 1 for each "provincia") (character);
\item \code{cod_provincia} codes of the Italian "province"; note that \code{cod_comune} is \code{cod_provincia} pasted to \code{progr_comune} (character);
\item \code{provincia} names of the Italian "province" (character);
\item \code{sigla_auto} automotive initials ("targa automobilistica") of the Italian "province" (character);
\item \code{cod_citta_metro} codes of the Italian "città metropolitane" (character);
\item \code{citta_metro} names of the Italian "città metropolitane" (character);
\item \code{cod_regione} codes of the Italian "regioni" (character);
\item \code{regione} names of the Italian "regioni" (character);
\item \code{cod_rip_geo} codes of the Italian "ripartizioni geografiche" (character);
\item \code{rip_geo} names of the Italian "ripartizioni geografiche" (character).
}
}

