% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{italy}
\alias{italy}
\title{Longitude and Latitude about italian cities}
\format{\code{italy} is a tbl data frame with 56 observations on 5 variables.}
\source{
\url{http://simplemaps.com/resources/world-cities-data}
}
\usage{
data(italy)
}
\description{
\code{italy} dataset contains information about latitude, longitude, population and region of the most important italian cities.
}
\details{
The 5 variables of the \code{italy} tbl data frame are the following ones:
\itemize{
 \item \code{city} city name (character)
 \item \code{lat} latitude (numeric)
 \item \code{lon} longitude (numeric)
 \item \code{pop} population (numeric)
 \item \code{region} region (character)
}
}

