% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{janka}
\alias{janka}
\title{Janka Hardness}
\format{\code{janka} is a tbl data frame with 36 observations on 2 variables.}
\usage{
data(janka)
}
\description{
Janka Hardness is an importance rating of Australian hardwood timbers. 
The test itself measures the force required to imbed a steel ball into a piece of wood and 
therefore provides a good indication to how the timber will withstand denting and wear. 
The dataset consists of density and hardness measurements from 36 Australian Eucalypt hardwoods.
}
\details{
The 2 variables of the \code{janka} tbl data frame are the following ones:
\itemize{
\item \code{Density} density measurements (numeric)
\item \code{Hardness} janka hardness (numeric)
}
}

