% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{knock}
\alias{knock}
\title{Knocking of the engine}
\format{\code{knock} is a tbl data frame with 13 observations on 5 variables.}
\usage{
data(knock)
}
\description{
The engeneers want to reduce the knocking of the engines. Before doing this, 
they have to identify which variables included in \code{knock} dataset influence this phenomenon. 
Data are randomly collected from 13 engines.
}
\details{
The 5 variables of the \code{knock} tbl data frame are the following ones:
\itemize{
\item \code{Spark} time of advance of the spark plug ignition (numeric)
\item \code{AFR} air fuel ratio (Air Fuel Ratio) (numeric)
\item \code{Intake} inlet temperature (numeric)
\item \code{Exhaust} exhaust temperature (numeric)
\item \code{Knock} knocking of the engine (numeric)
}
}

