% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{labonline}
\alias{labonline}
\title{pH measurement tools}
\format{\code{labonline} is a tbl data frame with 20 observations on 2 variables.}
\usage{
data(labonline)
}
\description{
Two pH measurement tools are compared: one instrument is a "gold standard", and is the "laboratory" tool, the other is an on-line (on the field) tool.
 Same samples are measured by using both instruments to see if on-line tool performs as laboratory tool.
}
\details{
The 2 variables of the \code{labonline} tbl data frame are the following ones:
\itemize{
\item \code{Lab} measuments collected with the "laboratory" tool (numeric)
\item \code{Online} measuments collected with the on-line tool (numeric)
}
}

