% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{life}
\alias{life}
\title{Life Expectations}
\format{\code{life} is a tbl data frame with 31 observations on 9 variables.}
\source{
Everitt, B. and Hothorn, T., An Introduction to Applied Multivariate Analysis with R, Springer, 2011
}
\usage{
data(life)
}
\description{
The \code{life} data contains life expectations in years at birth, 25, 50 and 75 years for men and women in 31 Countries or regions.
}
\details{
The 9 variables of the \code{life} tbl data frame are the following ones:
\itemize{
\item \code{country} (character);
\item \code{m0} life expectation in years for men at birth (numeric);
\item \code{m25} life expectation in years for men at 25 (numeric);
\item \code{m50} life expectation in years for men at 50 (numeric);
\item \code{m75} life expectation in years for men at 75 (numeric);
\item \code{m0} life expectation in years for women at birth (numeric);
\item \code{m25} life expectation in years for women at 25 (numeric);
\item \code{m50} life expectation in years for women at 50 (numeric);
\item \code{m75} life expectation in years for women at 75 (numeric).
}
}

