% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{mds}
\alias{mds}
\title{Fictional Data for MDS}
\format{\code{mds} is a tbl data frame with 10 observations on 5 variables.}
\usage{
data(mds)
}
\description{
The \code{mds} dataset is fictional and contains 10 observations on which 5 numerical variables are measured.
}
\details{
The 5 variables of the \code{mds} tbl data frame are the following ones:
\itemize{
\item \code{V1} (numeric);
\item \code{V2} (numeric);
\item \code{V3} (numeric);
\item \code{V4} (numeric);
\item \code{V5} (numeric);
}
}

