% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{mortality}
\alias{mortality}
\title{Mortality in the most important american cities}
\format{\code{mortality} is a tbl data frame with 60 observations on 15 variables.}
\source{
Data was adapted from StatLib Website.
}
\usage{
data(mortality)
}
\description{
\code{mortality} dataset contains information related to the percentage of mortality.
}
\details{
The 15 variables of the \code{mortality} tbl data frame are the following ones:
\itemize{
\item \code{Rain} annual average rainfall (numeric)
\item \code{JanTemp} average temperatures in January (numeric)
\item \code{JulyTemp} average temperatures in July (numeric)
\item \code{PctOver65} percentage of the population over 65 years (numeric)
\item \code{HHSize} average size of housing (numeric)
\item \code{Education} years of education (numeric)
\item \code{PctHomesLiveable} percentage of “habitable” homes (numeric)
\item \code{PopDensity} density of population (numeric)
\item \code{PctLowIncome} percentage of low-income families (numeric)
\item \code{PctWhiteCollar} percentage of employees (numeric)
\item \code{Hydrocarbon} pollution level by hydrocarbons (numeric)
\item \code{NititeOxide} pollution level of nitrite oxide (numeric)
\item \code{SulphurDioxide} pollution level of sulfur dioxide (numeric)
\item \code{RelHum} annual average relative humidity at 1 PM (numeric)
\item \code{MortalityRate} mortality rate for 100000 people (numeric)
}
}

