% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests}
\format{\code{mtcars} is a tbl data frame with 32 observations on 5 variables.}
\usage{
data(mtcars)
}
\description{
The \code{mtcars} data was extracted from the analogous data frame of package \code{datasets}, included in the base \code{R} distribution.
}
\details{
The 5 variables of the \code{mtcars} tbl data frame are the following ones:
\itemize{
\item \code{car} car model (character);
\item \code{carb} number of carburetors (numeric);
\item \code{cyl} number of cylinders (numeric);
\item \code{mpg} miles/(US) gallon (numeric);
\item \code{disp} displacement (cu.in.) (numeric).
}
}
\seealso{
the complete \code{mtcars} help page is in the \code{datasets} package: \code{\link[datasets]{mtcars}}
}

