% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{nes96}
\alias{nes96}
\title{nes96 Dataset}
\format{\code{nes96} is a tbl data frame with 944 observations on 10 variables.}
\source{
Sapiro, Virginia, Steven J. Rosenstone, Donald R. Kinder, Warren E. Miller, and the National Election Studies. AMERICAN NATIONAL ELECTION STUDIES, 1992-1997: COMBINED FILE [Computer file]. 2nd ICPSR version. Ann Arbor, MI: University of Michigan, Center for Political Studies [producer], 1999. Ann Arbor, MI: Inter-university Consortium for Political and Social Research [distributor], 1999
}
\usage{
data(nes96)
}
\description{
10 variable subset of the 1996 American National Election Study. Missing values and "don't know" responses have been listwise deleted. Respondents expressing a voting preference other than Clinton or Dole have been removed.
}
\details{
The 10 variables of the \code{nes96} tbl data frame are the following ones:
\itemize{
\item \code{popul} population of respondent's location in 1000s of people (numeric).
\item \code{TVnews} ays in the past week spent watching news on TV (numeric);
\item \code{selfLR} Left-Right self-placement of respondent: an ordered factor with levels extremely liberal, extLib < liberal, Lib < slightly liberal, sliLib < moderate, Mod < slightly conservative, sliCon < conservative, Con < extremely conservative, extCon (factor).
\item \code{ClinLR} Left-Right placement of Bill Clinton (same scale as selfLR): an ordered factor with levels extLib < Lib < sliLib < Mod < sliCon < Con < extCon (factor).
\item \code{DoleLR} Left-Right placement of Bob Dole (same scale as selfLR): an ordered factor with levels extLib < Lib < sliLib < Mod < sliCon < Con < extCon (factor).
\item \code{PID} Party identification: an ordered factor with levels strong Democrat, strDem < weak Democrat, weakDem < independent Democrat, indDem < independent independentindind < indepedent Republican, indRep < waek Republican, weakRep < strong Republican, strRep (factor).
\item \code{age} Respondent's age in years (numeric).
\item \code{educ} Respondent's education: an ordered factor with levels 8 years or less, MS < high school dropout, HSdrop < high school diploma or GED, HS < some College, Coll < Community or junior College degree, CCdeg < BA degree, BAdeg < postgraduate degree, MAdeg (factor).
\item \code{income} Respondent's family income: an ordered factor with levels $3Kminus < $3K-$5K < $5K-$7K < $7K-$9K < $9K-$10K < $10K-$11K < $11K-$12K < $12K-$13K < $13K-$14K < $14K-$15K < $15K-$17K < $17K-$20K < $20K-$22K < $22K-$25K < $25K-$30K < $30K-$35K < $35K-$40K < $40K-$45K < $45K-$50K < $50K-$60K < $60K-$75K < $75K-$90K < $90K-$105K < $105Kplus (factor).
\item \code{vote} Expected vote in 1996 presidential election: a factor with levels Clinton and Dole (factor).
}
}

