% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{orange}
\alias{orange}
\title{Growth of Orange Trees}
\format{\code{orange} is a tbl data frame with 35 observations on 3 variables.}
\source{
The original \code{Orange} dataset is included in \code{Datasets} package.
}
\usage{
data(orange)
}
\description{
\code{orange} dataset contains information about the growth of 5 Orange Trees, 
according to their trunk circumferences.
}
\details{
The 3 variables of the \code{orange} tbl data frame are the following ones:
\itemize{
 \item \code{Tree} tree id (character)
 \item \code{age} age of the tree (numeric)
 \item \code{circumference} trunk circumferences (mm) (numeric)
}
}

