% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{oxidant}
\alias{oxidant}
\title{Air Pollution}
\format{\code{oxidant} is a tbl data frame with 30 observations on 6 variables.}
\usage{
data(oxidant)
}
\description{
\code{oxidant} dataset contains data on levels of an air pollutant, "Oxidant", 
together with levels of four meteorological variables recorded on 30 days during one summer.
}
\details{
The 6 variables of the \code{oxidant} tbl data frame are the following ones:
\itemize{
\item \code{day} day index (numeric)
\item \code{windspeed} measure of windspeed (numeric)
\item \code{temperature} measure of temperature (numeric)
\item \code{humidity} measure of humidity (numeric)
\item \code{insolation} measure of insolation (numeric)
\item \code{oxidant} measure of oxidant (numeric)
}
}

