% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{ozone}
\alias{ozone}
\title{Ozone}
\format{\code{ozone} is a tbl data frame with 330 observations on 9 variables.}
\usage{
data(ozone)
}
\description{
\code{ozone} contains ozone data from Breiman and Friedman, 1985
}
\details{
The 9 variables of the \code{ozone} tbl data frame are the following ones:
\itemize{
\item \code{O3} Ozone conc., ppm, at Sandbug AFB. (numeric)
\item \code{temp} Temperature F. (max). (numeric)
\item \code{ibh} Inversion base height, feet (numeric)
\item \code{Pres} Daggett pressure gradient (mm Hg) (numeric)
\item \code{Vis} Visibility (miles) (numeric)
\item \code{Hgt} Vandenburg 500 millibar height (m) (numeric)
\item \code{Hum} Humidity, percent (numeric)
\item \code{ibt} Inversion base temperature, degrees F. (numeric)
\item \code{Wind} Wind speed, mph (numeric)
}
}

