% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{paint}
\alias{paint}
\title{Impurities in paints}
\format{\code{paint} is a tbl data frame with 12 observations on 2 variables.}
\usage{
data(paint)
}
\description{
\code{paint} contains the results of an experiment which has the aim to determine the relation between the rate 
of agitation applied to the container of paint and the number of lumps present in the containers.
}
\details{
The 2 variables of the \code{paint} tbl data frame are the following ones:
\itemize{
\item \code{Stirrate} rate of agitation (revolution) applied to the container (rpm, revolutions per minute) (numeric)
\item \code{Impurity} number of impurities (lumps) present in the containers of paint (numeric)
}
}

