% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{pencap}
\alias{pencap}
\title{Multi-head machine}
\format{\code{pencap} is a tbl data frame with 320 observations on 2 variables.}
\usage{
data(pencap)
}
\description{
\code{pencap} dataset contains information about the width of caps for pens produced by 16 different heads.
}
\details{
The 2 variables of the \code{pencap} tbl data frame are the following ones:
\itemize{
\item \code{Cavity} number of head (factor)
\item \code{Width} measurement (in mm) (numeric)
}
}

