% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{pimaindiansdiabetes2}
\alias{pimaindiansdiabetes2}
\title{Diabetes of Indian People}
\format{\code{pimaindiansdiabetes2} is a tbl data frame with 768 observations on 9 variables.}
\source{
National Institute of Diabetes and Digestive and Kidney Diseases
}
\usage{
data(pimaindiansdiabetes2)
}
\description{
The \code{pimaindiansdiabetes2} data contain information about diabetes of indian people.
}
\details{
The 9 variables of the \code{pimaindiansdiabetes2} tbl data frame are the following ones:
\itemize{
\item \code{pregnant} Number of times pregnant (numeric);
\item \code{glucose} Plasma glucose concentration a 2 hours in an oral glucose tolerance test (numeric);
\item \code{pressure} Diastolic blood pressure (mm Hg) (numeric);
\item \code{triceps} Triceps skin fold thickness (mm) (numeric);
\item \code{insulin} 2-Hour serum insulin (mu U/ml)(numeric);
\item \code{mass} Body mass index (weight in kg/(height in m)^2) (numeric);
\item \code{pedigree} Diabetes pedigree function (numeric);
\item \code{age} Age (years) (numeric);
\item \code{diabetes} class variable: diabetic or not (factor with 2 levels: \code{neg} and \code{pos});
}
}

