% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{plastic}
\alias{plastic}
\title{Comparison of two suppliers of plastics}
\format{\code{plastic} is a tbl data frame with 20 observations on 2 variables.}
\usage{
data(plastic)
}
\description{
\code{plastic} contains measurement data of resistance for wafer plastic of two suppliers.
}
\details{
The 2 variables of the \code{plastic} tbl data frame are the following ones:
\itemize{
\item \code{SupplrA} resistance of plastic wafer provided by supplier A (N) (numeric)
\item \code{SupplrB} resistance of plastic wafer provided by supplier B (N) (numeric)
}
}

