% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{polyester}
\alias{polyester}
\title{Polyesterification study}
\format{\code{polyester} is a tbl data frame with 6 observations on 2 variables.}
\usage{
data(polyester)
}
\description{
In the study of the polyesterification of fatty acids with glycols, the effect of temperature (C)
on the percentage conversion of the esterification process was investigated. 
Data are the results of an experiment using a catalyst 0.0004 mole zinc chloride per 100 grams of fatty acid.
}
\details{
The 2 variables of the \code{polyester} tbl data frame are the following ones:
\itemize{
\item \code{temperature} temperature effect (numeric)
\item \code{conversion} percentage conversion of the esterification process (numeric)
}
}

