% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{prostate}
\alias{prostate}
\title{Prostate Cancer}
\format{\code{prostate} is a tbl data frame with 97 observations on 9 variables.}
\source{
Stamey, T.A., Kabalin, J.N., McNeal, J.E., Johnstone, I.M., Freiha, F., Redwine, E.A. and Yang, N. (1989)
Prostate specific antigen in the diagnosis and treatment of adenocarcinoma of the prostate: II. radical prostatectomy treated patients, Journal of Urology 141(5), 1076–1083.
}
\usage{
data(prostate)
}
\description{
The \code{prostate} data come from a study that examined the correlation between the level of prostate specific antigen and a number of clinical measures in men who were about to receive a radical prostatectomy
}
\details{
The 9 variables of the \code{prostate} tbl data frame are the following ones:
\itemize{
\item \code{lcavol} log(cancer volume) (numeric);
\item \code{lweight} log(prostate weight) (numeric);
\item \code{age} age (numeric);
\item \code{lbph} log(benign prostatic hyperplasia amount) (numeric);
\item \code{svi} seminal vesicle invasion (numeric);
\item \code{lcp} log(capsular penetration) (numeric);
\item \code{gleason} Gleason score (numeric);
\item \code{pgg45} percentage Gleason scores 4 or 5 (numeric);
\item \code{lpsa} log(prostate specific antigen) (numeric);
}
}

