% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{province}
\alias{province}
\title{Data on Italian Districts (Province)}
\format{\code{province} is a tbl data frame with 110 observations on 5 variables.}
\source{
Istat website: \href{http://www.istat.it/it/archivio/6789}{Istituto nazionale di statistica (Istat)}.
}
\usage{
data(province)
}
\description{
The \code{province} dataset is related with information on the 110 Italian province and on the 9 Italian metropolitan cities
("città metropolitane"), updated at 2016-01-01.
}
\details{
The 5 variables of the \code{italia} tbl data frame are the following ones:
\itemize{
\item \code{cod_provincia} codes of the Italian "province"; 
\item \code{provincia} names of the Italian "province" (character);
\item \code{sigla_auto} automotive initials ("targa automobilistica") of the Italian "province" (character);
\item \code{cod_citta_metro} codes of the Italian "città metropolitane" (character);
\item \code{citta_metro} names of the Italian "città metropolitane" (character);
}
}

