% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata.R
\docType{package}
\name{qdata}
\alias{qdata}
\alias{qdata-package}
\title{Quantide Datasets Collection}
\description{
This package contains several datasets used during \href{http://www.quantide.com}{Quantide} courses.
}
\examples{
data(bands)
data(bank)
data(italia)
data(wwiileaders)
}
\author{
Enrico Tonini \email{enrico.tonini@quantide.com} \cr
Nicola Sturaro \email{nicola.sturaro@quantide.com} \cr
Enrico Pegoraro \email{enrico.pegoraro@quantide.com} \cr
Veronica Giro \email{veronica.giro@quantide.com} 

Maintainer: Nicola Sturaro \email{nicola.sturaro@quantide.com}
}

