% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{quine}
\alias{quine}
\title{Absenteeism from School in Rural New South Wales}
\format{\code{quine} is a tbl data frame with 146 observations on 5 variables.}
\source{
he complete \code{quine} help page is in the \code{MASS} package: \code{\link[MASS]{quine}}
}
\usage{
data(quine)
}
\description{
\code{quine} contains information about children from Walgett, New South Wales, 
Australia, who were classified by Culture, Age, Sex and Learner status and the number of days absent 
from school in a particular school year.
}
\details{
The 5 variables of the \code{quine} tbl data frame are the following ones:
\itemize{
\item \code{Eth} ethnic background: Aboriginal or Not (factor with levels: \code{A} and \code{N})
\item \code{Sex} sex (factor with levels: \code{F} and \code{M})
\item \code{Age} age group (factor with levels: \code{F0}, \code{F1}, \code{F2} and \code{F3})
\item \code{Lrn} learner status (factor with levels: \code{(AL} (Average learner) and \code{SL} (Slow learner))
\item \code{Days} days absent from school in the year
}
}

