% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{rats}
\alias{rats}
\title{Effect of toxic agent}
\format{\code{rats} is a tbl data frame with 48 observations on 3 variables.}
\usage{
data(rats)
}
\description{
\code{rats} contains measurement data of time survival of a sample of rats
 to which was administered three types of poison and four types of treatment. Each
 combination poison-treatment is administered to four rats.
}
\details{
The 3 variables of the \code{rats} tbl data frame are the following ones:
\itemize{
\item \code{Time} survival time (hours) (numeric)
\item \code{Poison} poison type (factor with levels: \code{I}, \code{II} and \code{III})
\item \code{Treatment} type of treatment (factor with levels: \code{A}, \code{B}, \code{C} and \code{D})
}
}

