% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{reaction}
\alias{reaction}
\title{Reaction temperature}
\format{\code{reaction} is a vector with 10 observations.}
\usage{
data(reaction)
}
\description{
A chemical theory suggests that the temperature at which a certain chemical reaction occurs is 180 °C. 
\code{reaction} data contains the results of 10 independent measurements of temperature.
}

