% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{regioni}
\alias{regioni}
\title{Data on Italian Districts (Regioni)}
\format{\code{regioni} is a tbl data frame with 20 observations on 2 variables.}
\source{
Istat website: \href{http://www.istat.it/it/archivio/6789}{Istituto nazionale di statistica (Istat)}.
}
\usage{
data(regioni)
}
\description{
The \code{regioni} dataset is related with information on the 20 Italian "regioni", updated at 2016-01-01.
}
\details{
The 2 variables of the \code{regioni} tbl data frame are the following ones:
\itemize{
\item \code{cod_regione} codes of the Italian "regioni" (character);
\item \code{regione} names of the Italian "regioni" (character);
}
}

