% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{ripartizioni}
\alias{ripartizioni}
\title{Data on Italian Districts (Ripartizioni)}
\format{\code{ripartizioni} is a tbl data frame with 5 observations on 2 variables.}
\source{
Istat website: \href{http://www.istat.it/it/archivio/6789}{Istituto nazionale di statistica (Istat)}.
}
\usage{
data(ripartizioni)
}
\description{
The \code{ripartizioni} dataset is related with information on the 5 Italian "ripartizioni".
}
\details{
The 2 variables of the \code{italia} tbl data frame are the following ones:
\itemize{
\item \code{cod_rip_geo} codes of the Italian "ripartizioni geografiche" (character);
\item \code{rip_geo} names of the Italian "ripartizioni geografiche" (character).
}
}

