% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{skin}
\alias{skin}
\title{Skin cancer}
\format{\code{skin} is a tbl data frame with 12 observations on 3 variables.}
\usage{
data(skin)
}
\description{
400 patients have been classified by cancer type and body site in which it has been detected.
}
\details{
The 3 variables of the \code{skin} tbl data frame are the following ones:
\itemize{
\item \code{Counts} counts of patients (numeric)
\item \code{Type} cancer type (factor with four levels: \code{Melan}, \code{Super}, \code{Nodul} and \code{Indet})
\item \code{Site} body site of cancer (numeric)
}
}

