% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{smoke}
\alias{smoke}
\title{Smoking Habits of Different Employees in a Company}
\format{Tbl data frame 6 observations and 4 variables with row manes.}
\source{
Michael J. Greenacre, *Theory and applications of correspondence analysis*, London: Academic Press, 1984
}
\usage{
data(smoke)
}
\description{
The \code{smoke} dataset contain information about the smoking habits in several employees of a company. 
The data frame is actually a contingency table of smoking habit (None, Light, Medium, Heavy) Vs. level of employee 
(Senior Manager, Junior Manager, Senior Employee, Junior Employee, Secretary). Data from Greenacre (1984, p. 55)
}
\details{
The 4 variables are:
\itemize{
\item \code{None} Non smokers counts (numeric)
\item \code{Light} Light smokers counts (character)
\item \code{Medium} Medium smokers counts (character)
\item \code{Heavy} Heavy smokers counts (character)
}
}

