% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{sres}
\alias{sres}
\title{Sres}
\format{\code{sres} is a tbl data frame with 19 observations on 3 variables.}
\usage{
data(sres)
}
\description{
The \code{sres} data was .....
}
\details{
The 3 variables of the \code{sres} tbl data frame are the following ones:
\itemize{
\item \code{V1} (numeric);
\item \code{V2} (numeric);
\item \code{V3} (numeric).
}
}

