% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{sturdy}
\alias{sturdy}
\title{Tear resistance}
\format{\code{sturdy} is a tbl data frame with 26 observations on 2 variables.}
\usage{
data(sturdy)
}
\description{
\code{sturdy} contains measurement data of tear resistance of five different raincoats brand.
}
\details{
The 2 variables of the \code{sturdy} tbl data frame are the following ones:
\itemize{
\item \code{Time} time between the solicitation and the tear (minutes) (numeric)
\item \code{Groups} raincoats brand (factor with levels: \code{A}, \code{B}, \code{C} and \code{D})
}
}

