% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{switcht}
\alias{switcht}
\title{Pressure Switch}
\format{\code{switcht} is a tbl data frame with 25 observations on 2 variables.}
\usage{
data(switcht)
}
\description{
\code{switcht} contains the measures of the pression (KPa) applied to 25 switches with different thickness of the membrane 
necessary for their open.
}
\details{
The 2 variables of the \code{switcht} tbl data frame are the following ones:
\itemize{
\item \code{DThickness} indicates the thickness of the diaphragm (mm) (numeric)
\item \code{SetPoint} indicates the pressure at which the switch opens (KPa) (numeric)
}
}

