% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{tissue}
\alias{tissue}
\title{Car seat tissues}
\format{\code{tissue} is a tbl data frame with 75 observations on 2 variables.}
\usage{
data(tissue)
}
\description{
\code{tissue} data contains the measures of the strenght of car seat tissues made by three quality inspectors of a plant: Henry, Anne, and Andrew (operators).
 Each operator measured 25 pieces of car seat tissues. Globally, 75 samples of tissue randomly chosen from the same production batch were measured.
}
\details{
The 2 variables of the \code{tissue} tbl data frame are the following ones:
\itemize{
\item \code{Operator} quality inspector name which measures the car seat tissue (factor with levels: \code{Anne}, \code{Henry} and \code{Andrew})
\item \code{Strength} measure of strength of the car seat tissue (numeric)
}
}

