% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{titanic}
\alias{titanic}
\title{Analysis of the data of the Titanic passengers}
\format{\code{titanic} is a tbl data frame with 2201 observations on 4 variables.}
\usage{
data(titanic)
}
\description{
\code{titanic} contains data concerning personal and travel informations of the single passengers, 
as well as their survival.
}
\details{
The 4 variables of the \code{titanic} tbl data frame are the following ones:
\itemize{
\item \code{Class} class of travel (factor with two levels: \code{First} and \code{Coach})
\item \code{Gender} gender of passengers (factor with two levels: \code{Male} and \code{Female})
\item \code{Age} age of passengers (numeric)
\item \code{Status} status of passengers (factor with two levels: \code{Died} and \code{Survived})
}
}

