% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{toxo}
\alias{toxo}
\title{Toxoplasmosis}
\format{\code{toxo} is a tbl data frame with 34 observations on 5 variables.}
\usage{
data(toxo)
}
\description{
The low birthweight data contains information on low birth weight in infants born at a US hospital. 
A number of variables were collected that might explain the cause of the low birth weight.
}
\details{
The 7 variables of the \code{banknotes} tbl data frame are the following ones:
\itemize{
\item \code{city}  city number (numeric)
\item \code{rain}  yearly rain (mm) (numeric)
\item \code{prop}  proportion of people affected by toxoplasmosis (numeric)
\item \code{num} examined people (numeric)
\item \code{ill} two-column matrix containing the number of non-illness (illN) and the number of illness (illY) (numeric matrix)
}
}

