% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{uscompanies}
\alias{uscompanies}
\title{79 Companies in the U.S.}
\format{\code{uscompanies} is a tbl data frame with 79 observations on 9 variables.}
\source{
Härdle and Simar, Applied Multivariate Statistical Analysis, 3rd ed., Springer, 2012
}
\usage{
data(uscompanies)
}
\description{
The \code{uscompanies} data deals with 79 U.S. companies classified by 6 numerical characteristics (assets, sales, etc.)
and one categorical characteristic (industry); indshort is a shortening for industry.
}
\details{
The 9 variables of the \code{uscompanies} tbl data frame are the following ones:
\itemize{
\item \code{company} car model (Factor);
\item \code{assets} (USD) (integer);
\item \code{sales} (USD) (integer);
\item \code{value} market value (USD) (integer);
\item \code{profits} (USD) (numeric);
\item \code{cashflow} (USD) (numeric);
\item \code{employees} (numeric);
\item \code{industry} (Factor);
\item \code{indshort} industry in a less fine classification (Factor);
}
}

