% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{uscrime}
\alias{uscrime}
\title{Crime Rates in the U.S.}
\format{\code{uscrime} is a tbl data frame with 50 observations on 12 variables.}
\source{
Härdle, W. K. and Simar, L., Applied Multivariate Statistical Analysis, 3rd edition, Springer, 2012
}
\usage{
data(uscrime)
}
\description{
The \code{uscrime} data deals with crime rates (number of crimes per 100,000 residents) in the 50 U.S.;
the dataset includes also land and population (in 1985) of the 50 U.S., which are classified by the following seven categories:
murder, rape, robbery, assault, burglary, larceny and auto-theft.
}
\details{
The 12 variables of the \code{uscrime} tbl data frame are the following ones:
\itemize{
\item \code{state} Usps states abbreviation (character);
\item \code{land} land area (integer);
\item \code{popu1985} population in 1985 (integer);
\item \code{murd} number of murders per 100,000 residents (numeric);
\item \code{rape} number of rapes per 100,000 residents (numeric);
\item \code{robb} number of robberies per 100,000 residents (numeric);
\item \code{assa} number of assaults per 100,000 residents (integer);
\item \code{burg} number of burglaries per 100,000 residents (integer);
\item \code{larc} number of larcenies per 100,000 residents (integer);
\item \code{auto} number of auto-thefts per 100,000 residents (integer);
\item \code{reg} U.S. states region number (integer);
\item \code{div} U.S. states division number (integer).
}
}

