% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{utilities}
\alias{utilities}
\title{22 Companies in the U.S.}
\format{\code{utilities} is a tbl data frame with 22 observations on 10 variables.}
\usage{
data(utilities)
}
\description{
The \code{utilities} data deals with 22 U.S. companies classified by 8 numerical characteristics (coverage, return, cost, etc.).
}
\details{
The 10 variables of the \code{utilities} tbl data frame are the following ones:
\itemize{

\item \code{coverage} fixed-charge coverage ratio (income/debt) (numeric);
\item \code{return} rate of return on capital (numeric);
\item \code{cost} cost per kW capacity in place (integer);
\item \code{load} annual load factor (numeric);
\item \code{peak} peak kWh demand growth from 1974 to 1975 (numeric);
\item \code{sales} sales (kWh use per year) (integer);
\item \code{nuclear} percent nuclear (numeric);
\item \code{fuel} total fuel costs (cents per kWh) (numeric);
\item \code{company} company full name (character);
\item \code{comp_short} company short name (character).
}
}

