% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{volcano3d}
\alias{volcano3d}
\title{Topographic Information on Auckland's Maunga Whau Volcano}
\format{\code{volcano3d} is a tbl data frame with 5307 observations on 3 variables.}
\usage{
data(volcano3d)
}
\description{
The \code{volcano3d} data was derived from the \code{volcano} data frame included in the base \code{R} distribution,
and gives the topographic information for Maunga Whau, one of about 50 volcanos in the Auckland volcanic field, on a 10m by 10m grid.
}
\details{
The 3 variables of the \code{volcano3d} tbl data frame are the following ones:
\itemize{
\item \code{x} x position (numeric)
\item \code{y} y position (numeric)
\item \code{z} z position (numeric)
}
}
\seealso{
the \code{volcano} help page in the \code{datasets} package: \code{\link[datasets]{volcano}}
}

