% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{wea}
\alias{wea}
\title{Weather Daily Data in Canberra}
\format{\code{wea} is a tbl data frame with 366 observations on 24 variables.}
\usage{
data(wea)
}
\description{
The \code{wea} data deals with weather daily data (temperature, rain, wind, etc.) in Canberra, Australia,
from 2007-11-01 to 2008-10-31.
}
\details{
The 5 variables of the \code{wea} tbl data frame are the following ones:
\itemize{
\item \code{Date} (Date), format: "2007-11-01" "2007-11-02" "2007-11-03" "2007-11-04" ...
\item \code{Location} Factor w/ 49 levels
\item \code{MinTemp} (numeric)
\item \code{MaxTemp} (numeric)
\item \code{Rainfall} (numeric)
\item \code{Evaporation} (numeric)
\item \code{Sunshine} (numeric)
\item \code{WindGustDir} Ord.factor w/ 16 levels
\item \code{WindGustSpeed} (numeric)
\item \code{WindDir9am} Ord.factor w/ 16 levels
\item \code{WindDir3pm} Ord.factor w/ 16 levels
\item \code{WindSpeed9am} (numeric)
\item \code{WindSpeed3pm} (numeric)
\item \code{Humidity9am} (numeric)
\item \code{Humidity3pm} (numeric)
\item \code{Pressure9am} (numeric)
\item \code{Pressure3pm} (numeric)
\item \code{Cloud9am} (numeric)
\item \code{Cloud3pm} (numeric)
\item \code{Temp9am} (numeric)
\item \code{Temp3pm} (numeric)
\item \code{RainToday} Factor w/ 2 levels
\item \code{RISK_MM} (numeric)
\item \code{RainTomorrow} Factor w/ 2 levels
}
}

