% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{wimbledon}
\alias{wimbledon}
\title{Open era: Wimbledon tournaments}
\format{\code{wimbledon} is a tbl data frames with 48 observations and 10 variables.}
\usage{
data(wimbledon)
}
\description{
The \code{wimbledon} datasets contain information about the gentlemen Wimbledon finals from 1968 to 2015.
}
\details{
The 10 variables of the \code{wimbledon} tbl data frame are the following ones:
\itemize{
\item \code{year} Tournament year (numeric)
\item \code{champion_country} champion's country (character)
\item \code{champion} champion's name and surname (character)
\item \code{runner_up_country} runner's country (character)
\item \code{runner_up} runner's name and surname (character)
\item \code{s1,s2,s3,s4,s5} scores in the five sets (character)
}
}

