% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{wines}
\alias{wines}
\title{Italian Wines from Piedmont, Italy}
\format{\code{wines} is a tbl data frame with 177 observations on 13 variables.}
\usage{
data(wines)
}
\description{
The \code{wines} data was extracted from the analogous data frame of package \code{kohonen} and contains information
on a set of 177 Italian wine samples from three different grape cultivars; 13 variables have been measured (alcohol, malic acid, tc.).
}
\details{
The 13 variables of the \code{wines} tbl data frame are the following ones:
\itemize{
\item \code{alcohol} (numeric)
\item \code{malic.acid} (numeric)
\item \code{ash} (numeric)
\item \code{ash.alkalinity} (numeric)
\item \code{magnesium} (numeric)
\item \code{tot..phenols} (numeric)
\item \code{flavonoids} (numeric)
\item \code{non.flav..phenols} (numeric)
\item \code{proanth} (numeric)
\item \code{col..int.} (numeric)
\item \code{col..hue} (numeric)
\item \code{OD.ratio} (numeric)
\item \code{proline} (numeric)
\item \code{vintages} Factor w/ 3 levels
\item \code{wine.classes} (numeric)
}
}
\seealso{
the complete \code{wines} help page in the \code{kohonen} package: \code{\link[kohonen]{wines}}
}

