% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{wwiileaders}
\alias{wwiileaders}
\title{Distance Matrix between 12 Country Leaders of II World War}
\format{\code{wwiileaders} is a distance matrix.}
\source{
Everitt, B. and Hothorn, T., *An Introduction to Applied Multivariate Analysis with `R`*, Springer, 2011
}
\usage{
data(wwiileaders)
}
\description{
The \code{wwiileaders} data contains a distance matrix among 12 country leaders during II World War,
which is based on the judgments of the dissimilarities in ideology. The subjects made judgments on a nine-point scale,
with the extreme points of the scale, 1 and 9, being described as indicating "very similar" and "very dissimilar", respectively.
}

