% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{people}
\alias{people}
\title{Weight and Height of 300 Italian People}
\format{\code{people} is a tbl data frame with 300 observations on 4 variables.}
\usage{
data(people)
}
\description{
The \code{people} dataset is related with information on 300 Italian people: their gender, area, weight and height.
}
\details{
The 4 variables of the \code{people} tbl data frame are the following ones:
\itemize{
\item \code{Gender} gender (factor with 2 levels: Female, Male);
\item \code{Area} residence area (factor with 4 levels: Centro, Isole, Nord, Sud);
\item \code{Weight} weight in kg (numeric);
\item \code{Height} height in cm (numeric).
}
}

