% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata-data.R
\name{uscrime}
\alias{uscrime}
\title{Crime Rates in the U.S.}
\format{\code{uscrime} is a tbl data frame with 50 observations on 12 variables.}
\source{
Härdle, W. K. and Simar, L., Applied Multivariate Statistical Analysis, 3rd edition, Springer, 2012
}
\usage{
data(uscrime)
}
\description{
The \code{uscrime} data deals with crime rates (number of crimes per 100,000 residents) in the 50 U.S.;
the dataset includes also land and population (in 1985) of the 50 U.S., which are classified by the following seven categories:
murder, rape, robbery, assault, burglary, larceny and auto-theft.
}
\details{
The 12 variables of the \code{uscrime} tbl data frame are the following ones:
\itemize{
\item \code{state} car model (character);
\item \code{land} land (integer);
\item \code{popu1985} population in 1985 (integer);
\item \code{murd} murder (numeric);
\item \code{rape} rape (numeric);
\item \code{robb} rabbery (numeric);
\item \code{assa} assault (integer);
\item \code{burg} burglary (integer);
\item \code{larc} larceny (integer);
\item \code{auto} auto-theft (integer);
\item \code{reg} ??? (integer);
\item \code{div} ??? (integer).
}
}

